Refactor ItemCategory to correct folder as per package diagram
