+++ CoffeeShop/src/main/java/item/ItemCategory.java
@@ -1,4 +1,4 @@
-package utils;
+package item;
 
 /**
  * Author: Cameron Hunt
